title: merge openact.c into open.c (single TU)
Folds openact.c back into open.c and drops its linker entry, so open+openact
compile to one object. gcc emits rodata in source order, so the single-TU byte
layout is identical to the two concatenated objects — title.bin still matches.
This removes the cross-object jump-table constraint (a function's switch table
and its code are always in the same object), unblocking title_open_800D2460
whose table lives at 0x800D8F00 while its code is at 0x800D2460.
